improved staking helper
